requirements.txt
streamlit
pandas
numpy
plotly
plotly.express
matplotlib.pyplot
seaborn